The default, fork-safe Boto3 session

The default, fork-safe Boto3 session

botor(
  aws_access_key_id,
  aws_secret_access_key,
  aws_session_token,
  region_name,
  botocore_session,
  profile_name
)

Arguments

aws_access_key_id

AWS access key ID

aws_secret_access_key

AWS secret access key

aws_session_token

AWS temporary session token

region_name

Default region when creating new connections

botocore_session

Use this Botocore session instead of creating a new default one

profile_name

The name of a profile to use. If not given, then the default profile is used

Value

boto3 Session